requirements.txt
Flask==3.0.3
GitPython==3.1.43
juju==3.4.0.0
pyyaml==6.0.1
celery==5.4.0
